s:MatchString(): Fix line continuation problem
I had had trouble with the line continuation in this function, but it
turns out what I had written the first time was fine. It's just that I
was sourcing my plugin incorrectly and that's the reason why I was
getting errors.

Before I was doing:

    $ vim --cmd 'so ~/Documents/Development/vim-tabs-grep/plugin/tabs_grep.vim' -p bin/imagebin.ca.sh bin/fucking-orange.sh functions/frequency w3m-tabs.exp

Using the `-c` flag instead completely fixed the problem:

    $ vim -c 'so ~/Documents/Development/vim-tabs-grep/plugin/tabs_grep.vim' -p bin/imagebin.ca.sh bin/fucking-orange.sh functions/frequency w3m-tabs.exp

Revert back to the original code in this function.