s:FilterTabPageElements(): Filter empty "Tab page X" elements
This function now works the way I want it to, removing the "Tab page X"
headers with no files under them from the `:tabs` list.

We do this by building a list of indexes of those "Tab page X" elements
and then removing them from the original list. The removal happens in
reverse so we don't mess up the indexes by removing from left to right.