Take a glob instead of a regex pattern
Globs make it easier to match file and directory names. For example,
instead of writing:

    :TabsGrep utol.*grep

we can write:

    :TabsGrep utol*grep

Discovered `glob2regpat()` in the help while searching for a way to take
a glob in my command and it ended up being the perfect thing, yay.

Needed to slice the regex returned by the function to discard the first
and last characters, as it automatically adds `^$` delimiters around the
generated regex pattern. This doesn't make sense for our use case
because we want to be able to start writing from anywhere in the middle
of the path, and not have to add `*`s to the beginning and end.